+++ Sources/RichTextKit/Images/Image+RichText.swift
@@ -11,6 +11,7 @@ import SwiftUI
 public extension Image {
 
     static let richTextCopy = symbol("doc.on.clipboard")
+    static let richTextDelete = symbol("trash")
     static let richTextDismissKeyboard = symbol("keyboard.chevron.compact.down")
     static let richTextEdit = symbol("square.and.pencil")
     static let richTextExport = symbol("square.and.arrow.up.on.square")